+++ NEWS.md
@@ -1,5 +1,10 @@
 # ggplot2 2.1.0.9000 
 
+* `layer()` gains new `check.aes` and `check.param` arguments. These allow
+  geom/stat authors to optional suppress checks for known aesthetics/parameters.
+  Currently this is used only in `geom_blank()` which powers `expand_limits()` 
+  (#1795).
+
 * Discrete scales should now work better with unicode labels on Windows (#1827).
 
 * When creating a layer, ggplot2 will warn if you use an unknown aesthetic